python/requirements.txt
keras==2.2.4
tensorflow==1.13.1
sympy==1.4
pillow==6.0.0
scikit-image==0.15.0
numpy==1.16.4